Implement sys/stat.h and define POSIX types with primitives to simplify everything

+++ src/stdio/_read.c
@@ -2,11 +2,11 @@
 
 extern _Thread_local int errno;
 
-ptrdiff_t __read(int, void*, size_t);
+long __read(int, void*, size_t);
 
 size_t __stdio_read(void* restrict buffer, size_t size, FILE stream[restrict static 1])
 {
-    ptrdiff_t count = __read(stream->fd, buffer, size);
+    long count = __read(stream->fd, buffer, size);
 
     if (count < 0) {
         stream->state |= _errbit;
